refactor: fix references to visuals
